@@ -5,28 +5,27 @@
   "main": "assets/javascripts/script.js",
   "dependencies": {
     "@babel/preset-react": "^7.6.3",
-    "babel-core": "^6.7.7",
-    "moment": "^2.29.2",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-moment": "^1.0.0",
-    "react-router": "^5.1.2",
-    "react-scripts": "^3.4.1",
-    "react-select": "^3.1.0",
-    "react-spinners": "^0.9.0",
+    "moment": "^2.29.3",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-moment": "^1.1.2",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "react-select": "^5.3.1",
+    "react-spinners": "^0.12.0-alpha.3",
     "recharts": "^2.1.9"
   },
   "devDependencies": {
-    "@babel/core": "7.11.6",
-    "@babel/plugin-proposal-class-properties": "7.10.4",
-    "@babel/preset-env": "7.11.5",
+    "@babel/core": "^7.17.10",
+    "@babel/plugin-proposal-class-properties": "^7.16.7",
+    "@babel/preset-env": "^7.17.10",
+    "@emotion/core": "^11.0.0",
     "babel-preset-react": "6.24.1",
-    "core-js": "3.6.5",
-    "react-dom": "16.13.1",
-    "react-router-dom": "5.2.0",
-    "webpack": "4.44.2",
-    "webpack-cli": "3.3.11",
-    "webpack-manifest-plugin": "2.2.0"
+    "core-js": "^3.22.4",
+    "webpack": "<5.53.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-manifest-plugin": "^5.0.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
