@@ -22,7 +22,7 @@ module.exports = {
               ['@babel/preset-react'],
               ['@babel/preset-env', { useBuiltIns: 'usage', corejs: 3 }]
             ],
-            plugins: ['@babel/plugin-proposal-class-properties', 'emotion']
+            plugins: ['@babel/plugin-proposal-class-properties', '@emotion']
           }
         }
       }
